DecomposingPolynomialsSystems.jl

The best package for simplifying systems of polynomial equations.

Package Features

Computing symmetries that fix the parameters of the polynomial system

Function Documentation

DecomposingPolynomialSystems.symmetries_fixing_parameters — Function

symmetries_fixing_parameters(F; degree)

This is an example of Docstring. This function returns the group of symmetries of the polynomial system F that fix the parameters.

\[Deck(f)\]