Monomials

Multiexponents

DecomposingPolynomialSystems.multiexponents — Function

multiexponents(; degree::T, nvars::S) where {T<:Integer,S<:Integer}

Generates vector of multiexponents of total degree degree each represented by SparseVector{T,S} of length nvars.

DenseMonomialVector

DecomposingPolynomialSystems.extend! — Function

extend!(mons::DenseMonomialVector{T,S}; degree::Integer) where {T<:Integer,S<:Integer}

Extends monomial vector mons by new multiexponents of total degree upto degree.

SparseMonomialVector

EvaluatedMonomials