Symmetries

Some decompositions are caused by the presence of non-trivial symmetries of the polynomial system that are given by rational maps. There are 2 important types of symmetries. The 1st type is scaling symmetries, namely, the maps that act by scaling individual variables of the formuation. The 2nd type are deck transformations, namely, the birational maps that fix the parameters of the polynomial system.

Scaling symmetries

DecomposingPolynomialSystems.scaling_symmetries — Function

scaling_symmetries(F::System; in_hnf::Bool=true)

Given a polynomial system F returns the group of scaling symmetries of the polynomial system F that fix the parameters.

julia> @var x[1:2] p[1:2]
(Variable[x₁, x₂], Variable[p₁, p₂])

julia> F = System([x[1]^2 - x[2]^2 - p[1], 2*x[1]*x[2] - p[2]]; variables=x, parameters=p)
System of length 2
 2 variables: x₁, x₂
 2 parameters: p₁, p₂

 -p₁ + x₁^2 - x₂^2
 -p₂ + 2*x₂*x₁

julia> scalings = scaling_symmetries(F)

Deck transformations

DecomposingPolynomialSystems.symmetries_fixing_parameters — Function

symmetries_fixing_parameters(F::System; degree_bound=1, param_dep=true, tol=1e-5)

Given a polynomial system F returns the group of symmetries of the polynomial system F that fix the parameters. The keyword argument degree_bound is used to set the upper bound for the degrees of numerator and denominator polynomials in expressions for the symmetries.

julia> @var x[1:2] p[1:2]
(Variable[x₁, x₂], Variable[p₁, p₂])

julia> F = System([x[1]^2 - x[2]^2 - p[1], 2*x[1]*x[2] - p[2]]; variables=x, parameters=p)
System of length 2
 2 variables: x₁, x₂
 2 parameters: p₁, p₂

 -p₁ + x₁^2 - x₂^2
 -p₂ + 2*x₂*x₁

julia> deck = symmetries_fixing_parameters(F; param_dep=false)